+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## _(in-progress)_
 * Unit tests:
     * New: Add test to verify plugin hooks `plugins_loaded` action to initialize itself
+    * New: Add tests for filters: `c2c_commenter_emails_filename`, `c2c_commenter_emails_show_emails`
 * Change: Note compatibility through WP 5.3+
 * Change: Update copyright date (2020)